+++ frontend/src/pages/RegionalDashboard/constants.js
@@ -1,0 +1,30 @@
+/* eslint-disable import/prefer-default-export */
+import {
+  regionFilter,
+  startDateFilter,
+  grantNumberFilter,
+  programSpecialistFilter,
+  programTypeFilter,
+  reasonsFilter,
+  recipientFilter,
+  reportIdFilter,
+  specialistRoleFilter,
+  stateCodeFilter,
+  targetPopulationsFilter,
+  topicsFilter,
+} from '../../components/filter/activityReportFilters';
+
+export const DASHBOARD_FILTER_CONFIG = [
+  startDateFilter,
+  grantNumberFilter,
+  programSpecialistFilter,
+  programTypeFilter,
+  reasonsFilter,
+  recipientFilter,
+  regionFilter,
+  reportIdFilter,
+  specialistRoleFilter,
+  stateCodeFilter,
+  targetPopulationsFilter,
+  topicsFilter,
+];
